refactor: move CSV hook and analysis components
